feat(android): port the beat editor — meter groups, undo, pads and preview
The editor was the largest remaining piece of the web app: 551 lines of
intricate draft state. It is now ported as a pure, immutable EditorDraft
model plus a Compose screen that owns only presentation, so every rule is
unit-tested without a compositor.

- Three zones as on the web: an overview mini-strip with a draggable page
  frame, a zoomed eight-cell page whose cursor is a column, and one pad per
  enterable stroke that writes, sounds and advances in one gesture. Lane
  labels double as isolate toggles, which is how the cymbal row is authored.
- Meter model is a list of groups (cells per numbered beat), so uneven
  signatures like 7/8 = [2,2,3] are expressible; the feel picker seeds 4/4,
  12/8, 16ths, 6/8 and 7/8, and +/- grows the bar by one whole unit.
- Undo is a 60-deep stack of previous drafts; resizing or changing the feel
  resets it because old snapshots carry the old length.
- Saving persists the kartal lane only when the user drew one, keeps an
  existing custom beat's id so it overwrites in place, and stays open on a
  failed write — unmounting is how a draft is discarded, so closing on
  failure would throw the work away.

Supporting model change: Beat.beatsPerBar is now a Double. 7/8 at eighth
subdivision is 3.5 quarters to the bar and the web app stores exactly that;
an Int silently retimed every uneven meter the editor can author. The clock
turns PPQ x beatsPerBar into whole ticks, so 7/8 loops at 672 ticks exactly.
LibraryJson gained a fractional field reader so a 3.5 from the web or the
cloud is honoured rather than refused.

Verified: 268 unit tests (30 new for the draft model), and on the emulator
end to end — four pad taps wrote four dayan marks with the cursor advanced
and the other lanes untouched, save closed the editor and the beat appeared
under Your beats. Installed on the physical device as well.

+++ android/app/src/main/java/com/kirtan/companion/data/ShareCodec.kt
@@ -319,7 +319,7 @@ object ShareCodec {
             note = "Custom",
             bpm = bpm,
             steps = steps,
-            beatsPerBar = steps / q,
+            beatsPerBar = steps.toDouble() / q,
             cellsPerGroup = if (uniform) groups[0] else q,
             groups = groups,
             description = null,      // dropped, not validated; see rule 5
